requirements.txt
python-telegram-bot
openai
langdetect